fix a dangerous bug of quick_hist, it turns out that I didn't ensure that the weigths are np.float64. Before for example big endian float64 were completely misinterpreted

+++ my_utils/quick_hist.py
@@ -105,7 +105,7 @@ def quick_hist(arrs, range=None, nbins=None, weights=None, getPos=False):
 	if weights is None:
 		weights_str = '1'
 	else:
-		weightsind = np.ascontiguousarray(weights[ind])
+		weightsind = np.ascontiguousarray(weights[ind], dtype=np.float64)
 		weights_str = 'weightsind(i)'
 	if not getPos:	
 		del ind
